fix(tests): repair test + example compile targets blocking cargo test --workspace
Make `cargo test --workspace --no-run` (and all examples/benches) compile.

Test targets
- Add `newport_core` shim crate (tests/support/newport_core) providing the
  legacy byte-addressed Memory/TileId/RaceWayPacket/NewportError API the
  stress / test_utils / verilog_cross_validation / simulation_bench targets
  were written against (pre-Cognitum rename). Real impls, nothing stubbed.
- Wire `cognitum-security` (path) as a dev-dependency with a new `mock`
  feature so unit/acceptance security tests resolve `cognitum_security::*`.
  Gate `MockHsmProvider` generation/export on `any(test, feature = "mock")`
  via an optional mockall dep — production builds stay lean.
- Drop the incorrect `harness = false` on acceptance_tests/unit_tests
  (they use the default libtest harness; the override caused E0601 no-main).

Examples
- snn_router_benchmark.rs: `println!("-".repeat(50))` -> `println!("{}", ...)`
  (5 sites) — the bare method-call expr was a macro parse error.
- simulator_bridge_demo.rs: scope the immutable bridge borrows so they no
  longer overlap the mutable get_routing_bridge_mut() calls (E0502).

Source (real missing exports / test-support accessors, no logic changes)
- hipaa: re-export PhiField + AccessDecision from `hipaa`.
- audit: add AuditLogger::store() accessor.
- ruvector::fusion: add BrittlenessMonitor::min_cut_history() accessor.
- hipaa::storage: add HipaaCompliantStorage::get_encrypted() accessor.

Test fixes (resolve missing imports / borrow / mock-trait issues only)
- import AuditStore trait where `.append()/.query()` are used.
- import PartitionId in fusion_test.
- clone RuvectorConfig before reuse in routing_test.
- define real QuotaStore/LicenseStore traits backing the mock! blocks.
- widen tile-id range asserts to u16 (value() is u8; `< 256` overflowed).
- ruvector_bench: clone non-Copy TileState inside the FnMut closure.

+++ tests/stress_tests.rs
@@ -100,8 +100,8 @@ fn stress_test_packet_flood() {
     while let Some(packet) = queue.pop_front() {
         // Verify packet integrity
         assert!(!packet.data.is_empty());
-        assert!(packet.source.value() < 256);
-        assert!(packet.dest.value() < 256);
+        assert!((packet.source.value() as u16) < 256);
+        assert!((packet.dest.value() as u16) < 256);
         processed += 1;
     }
